feat: Improve SIMS pretraining loss function
The SIMS pretraining loss function has been updated to improve stability and performance. The new loss function, `UnsupervisedLossSimsational`, incorporates additional features such as adding epsilon to stabilize batch standard deviations and scaling the loss by the number of features. This change aims to address the issue of the loss not decreasing during training.